Implement Scheme reader labels; implement issue #2
This implements Scheme reader labels and fixes two small bugs. The
first small bug is that + and - were read as characters rather than as
symbols. The second "bug" was that a function called read-scheme-file
did not actually accept pathnames or path strings as input. This means
that issue #2 is finally complete.